Renaming function files for consistancy
